refactor: consolidate `SignBidirectionalEventExt` usage across modules
