+++ chain-signatures/node/src/backlog/mod.rs
@@ -3,16 +3,15 @@ pub mod selection;
 use self::selection::select_checkpoints;
 use crate::mesh::MeshState;
 use crate::node_client::NodeClient;
-use crate::sign_bidirectional::{PublishState, SignStatus};
+use crate::sign_bidirectional::{PublishState, SignBidirectionalEventExt, SignStatus};
 use crate::storage::checkpoint_storage::CheckpointStorage;
-use crate::stream::ops::SignBidirectionalEventExt;
 
 use anyhow::Context;
 use mpc_primitives::{
     BidirectionalTx, BidirectionalTxId, Chain, IndexedSignRequest, PendingTx, SignId, SignKind,
 };
 use sha3::{Digest, Sha3_256};
-use std::collections::{HashMap, hash_map};
+use std::collections::{hash_map, HashMap};
 use std::hash::{Hash, Hasher};
 use std::sync::Arc;
 use std::time::{Duration, Instant};
