+++ integration-tests/tests/cases/canton_stream.rs
@@ -6,8 +6,7 @@ use mpc_node::backlog::Backlog;
 use mpc_node::indexer_canton::contracts::{CantonSignature, EcdsaSigData};
 use mpc_node::indexer_canton::{der_encode_signature, CantonStream};
 use mpc_node::protocol::{Chain, IndexedSignRequest};
-use mpc_node::sign_bidirectional::hash_rlp_data;
-use mpc_node::stream::ops::SignBidirectionalEventExt;
+use mpc_node::sign_bidirectional::{hash_rlp_data, SignBidirectionalEventExt};
 use mpc_node::stream::{catchup_then_livestream, ChainStream};
 use mpc_primitives::{ChainEvent, ScalarExt, SignKind, Signature, LATEST_MPC_KEY_VERSION};
 use serde_json::json;
